==> notebooks/1.0-jtr-prepare-daily-schedule.ipynb <==
# Prepare a Daily Schedule
Before conducting analysis on a day's worth of delay data, there is a lot of data in the schedule datasets that can be removed to make our life a lot easier.

By the end of this notebook, we will have access to two tables: the trips, and the stop times for each trip. These tables will have columns ready to parse in delay information when this has been analysed.

## Accessing the archive
This project archives each day's schedule information.

import zipfile
import io
file_name = '20190124.zip'
with open(file_name, "rb") as f:
    z = zipfile.ZipFile(io.BytesIO(f.read()))

z.extractall()
print("Extracted file " + file_name)

This contains all the set schedules, as well as the real time delay information, which we will deal with later.

## The schedules
The schedules are a bunch of text files.

import glob
data_path = 'home/pi/sydney-transport-tracker/data/raw/20190124/'
timetable_files = glob.glob(data_path + '*.txt')
print('The timetable files are:\n' + '\n'.join(timetable_files))

This is the data you will get when querying the Transport for NSW Open Data [schedules API](https://opendata.transport.nsw.gov.au/dataset/public-transport-timetables-realtime). 

### Trips
`trips.txt` is every trip that will run over a period of time (more than 1 day). Trains running on the same "line" will share a `route_id`, each will have a unique `trip_id`, but they also have a `service_id` which can be used to determine which trips are running on what day.

These schedule text files that are downloaded may be relevant for a week or more. What we want to do is filter out this trips file so that it only contains the trips we are interested in, that ran on the 23/01/2019 (and on the city network only - we will get to that next).

To determine what trips ran today, we look at the `calendar.txt` file. It contains the `service_id` (matching those from `trips.txt`) for each day in this period.

import datetime
import csv

day_of_analysis = 'thursday'
date_of_analysis = datetime.datetime.strptime('20190124', "%Y%m%d").date()
todays_services = []

with open(data_path + 'calendar.txt', mode='r', encoding='utf-8-sig') as csv_file:
    csv_reader = csv.DictReader(csv_file)
    line_count = 0
    for row in csv_reader:
        if row[day_of_analysis] == '1':
            start_date = datetime.datetime.strptime(row['start_date'], "%Y%m%d").date()
            end_date = datetime.datetime.strptime(row['end_date'], "%Y%m%d").date()
            if start_date <= date_of_analysis <= end_date:
                todays_services.append(row['service_id'])

print("Todays services:\n" + '\n'.join(todays_services))

With this information, we can filter out all the trips from `trips.txt` that do not run on this day.

import pandas as pd
df_trips = pd.read_csv(data_path + 'trips.txt',
                       header=0,
                       encoding='utf-8-sig',
                       usecols=["route_id", "service_id", "trip_id"])
df_trips = df_trips[df_trips['service_id'].isin(todays_services)]
pd.options.display.max_rows = 10
df_trips

If you were peruse the entire dataset, you would find that there are some `route_id` values that are non-commuter trains (see page 4 of the [Sydney Trains Realtime GTFS & GTFS- R Technical Document](https://opendata.transport.nsw.gov.au/sites/default/files/Real-Time_Train_Technical_Document_v2.5.pdf)). As well as this, there are interstate and regional services that we don't really want to consider when analysing delays, as they are long journeys rather than a daily commute.
Let's filter those out.

ROUTES_TO_IGNORE = ["CTY_NC1", "CTY_NC1a", "CTY_NC2", 
                    "CTY_NW1a", "CTY_NW1b", "CTY_NW1c", "CTY_NW1d", "CTY_NW2a", "CTY_NW2b", 
                    "CTY_S1a", "CTY_S1b", "CTY_S1c", "CTY_S1d", "CTY_S1e", "CTY_S1f", 
                    "CTY_S1g", "CTY_S1h", "CTY_S1i", 
                    "CTY_S2a", "CTY_S2b", "CTY_S2c", "CTY_S2d", "CTY_S2e", "CTY_S2f", 
                    "CTY_S2g", "CTY_S2h", "CTY_S2i", 
                    "CTY_W1a", "CTY_W1b", "CTY_W2a", "CTY_W2b", 
                    "HUN_1a", "HUN_1b", "HUN_2a", "HUN_2b", 
                    "RTTA_DEF", "RTTA_REV"]
df_trips = df_trips[~df_trips['route_id'].isin(ROUTES_TO_IGNORE)]
df_trips

df_trips.to_pickle('trips.pickle')

That removed hundreds of trips from consideration.

### Stops
`stop_times.txt`, like `trips.txt`, contains information for journeys that occur on different days, and on routes we don't really care about.
Now that we have a table containing every `trip_id` under consideration, we can filter out all of the stop times that don't matter.

df_stop_times = pd.read_csv(data_path + 'stop_times.txt', header=0,
                            encoding='utf-8-sig',
                            dtype={'stop_id': str},
                            usecols=["trip_id", "arrival_time", "departure_time", "stop_id"],
                            parse_dates=['arrival_time', 'departure_time'])

# remove any trips from stop_times that did NOT happen on this date, using the trips dataset
df_stop_times = df_stop_times[df_stop_times['trip_id'].isin(df_trips['trip_id'])]
df_stop_times

df_stop_times.to_pickle('stop_times.pickle')

Now that we're only dealing with a subset of the original data dump from the Open Data API, we can start parsing in the real-time delay data in the next notebook.
