==> notebooks/3.0-jtr-merge-delays-schedules.ipynb <==
# Merging delays with the schedules
With the daily schedule formatted into two tables in the first notebook, and delay information summarised into a single object in the second notebook, it's now time to merge the information so that we can easily get insights into how the trains ran that day.

## Full schedule with delays
We want today's timetable, but with the delays on a per stop basis.
Import the dataset formatted from the first notebook.

import pandas as pd
pd.options.display.max_rows = 10
df_stop_times = pd.read_pickle('stop_times.pickle')
df_stop_times

Add columns for arrival and departure times and delays, and schedule relationship. Mark all as N/A to start with, so we know which things never had real time updates

df_stop_times.insert(2, 'arrival_delay', 'N/A')
df_stop_times.insert(3, 'actual_arrival_time', 'N/A')
df_stop_times.insert(5, 'departure_delay', 'N/A')
df_stop_times.insert(6, 'actual_departure_time', 'N/A')
df_stop_times.insert(7, 'schedule_relationship', 'N/A')
df_stop_times

Then import the single delays object, and merge the data. Be warned, this is a very time consuming operation.

import pickle
import sys
sys.path.append('../')
from src.features.trip_objects import *
from src.features.trip_helper import *

merged_delays = pickle.load(open('merged_delays.pickle', 'rb'))

# iterate through all the trips we 
for trip in merged_delays.values():
    df_stop_times_this_trip = df_stop_times[(df_stop_times['trip_id'] == trip.trip_id)]
    for stop_time_update in trip.stop_time_updates.values():
        # some of these values might be 24:00, 25:00 etc to signify next day

        idx = df_stop_times_this_trip[(df_stop_times_this_trip['stop_id'] == stop_time_update.stop_id)].index
        if idx.empty:
            # it shouldn't be
            continue
            
        idx = idx.item()

        # calculate the real time
        actual_arrival_time = update_time('20190124', df_stop_times_this_trip.at[idx, 'arrival_time'],
                                               stop_time_update.arrival_delay)
        actual_departure_time = update_time('20190124', df_stop_times_this_trip.at[idx, 'departure_time'],
                                                 stop_time_update.departure_delay)

        # add the new values to the new columns
        df_stop_times.at[idx, 'arrival_delay'] = stop_time_update.arrival_delay
        df_stop_times.at[idx, 'actual_arrival_time'] = actual_arrival_time
        df_stop_times.at[idx, 'departure_delay'] = stop_time_update.departure_delay
        df_stop_times.at[idx, 'actual_departure_time'] = actual_departure_time
        df_stop_times.at[idx, 'schedule_relationship'] = stop_time_update.schedule_relationship

df_stop_times

Let's see if we have any arrival delay statistics:

df_stop_times[df_stop_times['arrival_delay'] != 'N/A']

df_stop_times.to_pickle('stop_times_merged.pickle')

## Trips
Now import the trips table, and merge the insights from the delay object into it.

df_trips = pd.read_pickle('trips.pickle')
df_trips

Add columns for the start and end timestamps, the delay averages and maxima, and the schedule relationship of each trip.

df_trips.insert(0, 'start_timestamp', 'N/A')
df_trips.insert(1, 'end_timestamp', 'N/A')
df_trips.insert(5, 'maximum_arrival_delay', 0)
df_trips.insert(6, 'average_arrival_delay', 0)
df_trips.insert(7, 'maximum_departure_delay', 0)
df_trips.insert(8, 'average_departure_delay', 0)
df_trips.insert(9, 'schedule_relationship', 0)
df_trips

Add the insights for each trip to the corresponding row of the table.

for trip in merged_delays.values():

    idx = df_trips[(df_trips['trip_id'] == trip.trip_id)].index
    if idx.empty:
        # it shouldn't be
        continue

    idx = idx.item()

    df_trips.at[idx, 'maximum_arrival_delay'] = trip.maximum_arrival_delay()
    df_trips.at[idx, 'average_arrival_delay'] = trip.average_arrival_delay()
    df_trips.at[idx, 'maximum_departure_delay'] = trip.maximum_departure_delay()
    df_trips.at[idx, 'average_departure_delay'] = trip.average_departure_delay()
    df_trips.at[idx, 'schedule_relationship'] = trip.overall_schedule_relationship()

df_trips

For every trip in the table, add the full start and end timestamps. This makes it easier to deal with information when a trip runs from very late to very early the next day, and the dates change.

for i in df_trips.index:
    departure_series = df_stop_times[df_stop_times['trip_id'] == df_trips.at[i, 'trip_id']]['departure_time']
    if len(departure_series) < 2:
        continue
    df_trips.at[i, 'start_timestamp'] = convert_to_timestamp('20190124', departure_series.iloc[0])
    df_trips.at[i, 'end_timestamp'] = convert_to_timestamp('20190124', departure_series.iloc[-1])

df_trips

What were the worst delays?

df_trips.sort_values(by=['maximum_departure_delay'], ascending=False).head(n=10)

84 minutes!

df_trips.to_pickle('trips_insights.pickle')

import matplotlib.pyplot as plt
df_trips.plot(x="start_timestamp", y=["maximum_arrival_delay"])

df_trips_morning = df_trips[df_trips['start_timestamp'] > datetime.datetime(2019,1,24,6,0)]
pd.options.display.max_rows = 70
df_trips_morning[df_trips_morning['route_id'].str.contains("IWL")].sort_values(by=['start_timestamp']).head(n=70)
